requirements.txt
Click==7.0
Flask==1.1.1
Flask-SocketIO==4.2.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
peewee==3.10.0
python-engineio==3.9.3
python-socketio==4.3.1
six==1.12.0
speedtest-cli==2.1.1
Werkzeug==0.15.5
